fix(firecracker): compile process signals cross-platform
