+++ internal/providers/firecracker/process_unix.go
@@ -1,0 +1,26 @@
+//go:build !windows
+
+package firecracker
+
+import (
+	"errors"
+	"syscall"
+)
+
+func signalProcess(pid int, sig syscall.Signal) error {
+	if pid <= 0 {
+		return nil
+	}
+	if err := syscall.Kill(pid, sig); err != nil && !errors.Is(err, syscall.ESRCH) {
+		return err
+	}
+	return nil
+}
+
+func processAlive(pid int) bool {
+	if pid <= 0 {
+		return false
+	}
+	err := syscall.Kill(pid, 0)
+	return err == nil || errors.Is(err, syscall.EPERM)
+}
