+++ internal/providers/firecracker/process_windows.go
@@ -1,0 +1,13 @@
+//go:build windows
+
+package firecracker
+
+import "syscall"
+
+func signalProcess(int, syscall.Signal) error {
+	return nil
+}
+
+func processAlive(int) bool {
+	return false
+}
